refactor!: conservationrules -> conservation_rules
