+++ cspell.json
@@ -26,14 +26,18 @@
         "conda",
         "docstrings",
         "expertsystem",
+        "gell",
         "gordan",
         "graphviz",
         "helicity",
         "ipynb",
         "ipython",
         "isospin",
+        "itertools",
         "jupyter",
+        "mathbb",
         "mypy",
+        "nishijima",
         "numpy",
         "pbar",
         "pycompwa",
@@ -43,6 +47,7 @@
         "redeboer",
         "spflueger",
         "tensorwaves",
+        "toctree",
         "venv",
         "weisskopf",
         "xmltodict"
@@ -53,8 +58,12 @@
         "codacy",
         "codemirror",
         "cparity",
+        "genindex",
+        "gparity",
         "kernelspec",
+        "maxdepth",
         "mimetype",
+        "modindex",
         "nbconvert",
         "nbformat",
         "nbsphinx",
