+++ expertsystem/state/__init__.py
@@ -1,11 +1,11 @@
 """All modules related to a quantum mechanical state."""
 
 __all__ = [
-    "conservationrules",
+    "conservation_rules",
     "particle",
     "propagation",
 ]
 
-from . import conservationrules
+from . import conservation_rules
 from . import particle
 from . import propagation
