+++ tests/conservation_laws/test_gparity.py
@@ -1,4 +1,4 @@
-from expertsystem.state.conservationrules import GParityConservation
+from expertsystem.state.conservation_rules import GParityConservation
 from expertsystem.state.particle import (
     InteractionQuantumNumberNames,
     ParticlePropertyNames,
