+++ tests/conservation_laws/test_spin.py
@@ -1,4 +1,4 @@
-from expertsystem.state.conservationrules import SpinConservation
+from expertsystem.state.conservation_rules import SpinConservation
 from expertsystem.state.particle import (
     InteractionQuantumNumberNames,
     Spin,
